refactor(editor): extract Track D interactions and canvas render (PR 10.15–10.19)
Peel remaining high-coupling editor event families and the main canvas
draw path into behavior-owned modules while keeping wire_editor_events
and public setup entry points stable.

- interaction: SpacePanState + EyedropperBundle; reuse apply_zoom_change
- drag/click/motion/keyboard: complete gesture and key families
- canvas_render: CanvasRenderCaches + install_canvas_draw_func
- docs: validation/status updated for Track D completion
